refactor(pdf-service): move /domains/shared/pdf to /domains/shared/services/pdf
