+++ app/domains/beratungshilfe/services/pdf/pdfForm/F_besitz/F_besitz.ts
@@ -1,11 +1,11 @@
-import type { AttachmentEntries } from "~/domains/shared/pdf/attachment";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
+import type { AttachmentEntries } from "~/domains/shared/services/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
 import {
   attachBankkontenToAnhang,
   attachGrundeigentumToAnhang,
   eigentuemerMapping,
   grundeigentumArtMapping,
-} from "~/domains/shared/pdf/eigentumHelpers";
+} from "~/domains/shared/services/pdf/eigentumHelpers";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import { arrayIsNonEmpty } from "~/util/array";
 import { fillKraftfahrzeug } from "./fillKraftfahrzeug";
