+++ app/domains/beratungshilfe/services/pdf/pdfForm/F_besitz/__test__/F_besitz.test.ts
@@ -1,6 +1,6 @@
 import { getBeratungshilfeParameters } from "data/pdf/beratungshilfe/beratungshilfe.generated";
 import { type BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import {
   fillFinancialBankkonto,
