+++ app/domains/beratungshilfe/services/pdf/pdfForm/F_besitz/fillVermoegenswerte.ts
@@ -1,10 +1,10 @@
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
 import {
   attachGeldanlagenToAnhang,
   eigentuemerMapping,
   fillSingleGeldanlage,
   fillSingleWertsache,
-} from "~/domains/shared/pdf/eigentumHelpers";
+} from "~/domains/shared/services/pdf/eigentumHelpers";
 import type { BerHPdfFillFunction } from "../..";
 
 const VERMOEGENSWERT_BEZEICHNUNG_FIELD_MAX_CHARS = 148;
