+++ app/domains/prozesskostenhilfe/services/pdf/index.ts
@@ -2,7 +2,7 @@ import { PDFDocument } from "pdf-lib";
 import type { ProzesskostenhilfePDF } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import { getProzesskostenhilfeParameters } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import { createAttachmentPages } from "~/domains/shared/pdf/attachment/createAttachmentPages";
+import { createAttachmentPages } from "~/domains/shared/services/pdf/attachment/createAttachmentPages";
 import { appendPagesToPdf } from "~/services/pdf/appendPagesToPdf";
 import {
   type PdfFillFunction,
