+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/C_unterhaltspflichtige_person.ts
@@ -1,4 +1,4 @@
-import { familyRelationshipMap } from "~/domains/shared/pdf/unterhaltHelpers";
+import { familyRelationshipMap } from "~/domains/shared/services/pdf/unterhaltHelpers";
 import type { PkhPdfFillFunction } from "..";
 
 export const fillUnterhaltsanspruch: PkhPdfFillFunction = ({
