+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/J_belastungen.ts
@@ -1,6 +1,6 @@
 import { ausgabenSituationMapping } from "~/domains/beratungshilfe/services/pdf/pdfForm/G_ausgaben";
 import { finanzielleAngabeEinkuenfteGuards as einkuenfteGuards } from "~/domains/prozesskostenhilfe/formular/finanzielleAngaben/einkuenfte/guards";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/services/pdf/attachment";
 import type { PkhPdfFillFunction } from "..";
 
 export const fillBelastungen: PkhPdfFillFunction = ({
